Rename BackboneViewMixin to Events
