+++ test/unit/collection-view/collection-view.spec.js
@@ -5,7 +5,7 @@ import _ from 'underscore';
 import Backbone from 'backbone';
 import CollectionView from '../../../src/collection-view';
 import View from '../../../src/view';
-import BackboneViewMixin from '../../../src/mixins/backboneview';
+import Events from '../../../src/mixins/events';
 
 describe('CollectionView', function() {
   let MyChildView;
@@ -26,7 +26,7 @@ describe('CollectionView', function() {
       onBeforeDestroy: this.sinon.stub(),
       onDestroy: this.sinon.stub(),
     });
-    _.extend(MyBbChildView.prototype, BackboneViewMixin);
+    _.extend(MyBbChildView.prototype, Events);
   });
 
   describe('#constructor', function() {
@@ -150,7 +150,7 @@ describe('CollectionView', function() {
     describe('when childView is a Backbone.View', function() {
       it('should build children from the defined view', function() {
         let BBView = Backbone.View.extend();
-        _.extend(BBView.prototype, BackboneViewMixin);
+        _.extend(BBView.prototype, Events);
         const myCollectionView = new CollectionView({
           collection,
           childView: BBView
@@ -165,7 +165,7 @@ describe('CollectionView', function() {
       let myCollectionView;
       let childViewStub;
       let BBView = Backbone.View.extend();
-      _.extend(BBView.prototype, BackboneViewMixin);
+      _.extend(BBView.prototype, Events);
       beforeEach(function() {
         childViewStub = this.sinon.stub();
         childViewStub.returns(BBView);
